db_berkeley: Rename module interface file #82
