+++ src/modules/db_berkeley/db_berkeley.c
@@ -44,7 +44,7 @@
 #include "../../lib/srdb1/db_res.h"
 #include "../../lib/srdb1/db.h"
 #include "../../lib/srdb1/db_query.h"
-#include "km_db_berkeley.h"
+#include "db_berkeley.h"
 #include "km_bdb_lib.h"
 #include "km_bdb_res.h"
 #include "bdb_mod.h"
